Added submodule config [tested]
- Moved Configuration, ConfigurationError and OptimizierConfigurator from fluke.utils to fluke.config
