+++ docs/api_reference.md
@@ -10,6 +10,7 @@ In this page you can find the list of modules/submodules defined in `fluke` with
 - [`fluke.data`](fluke.data.md): contains classes for data handling;
 - [`fluke.data.datasets`](fluke.data.datasets.md): contains classes for datasets loading;
 - [`fluke.client`](fluke.client.md): contains classes for client-side functionalities;
+- [`fluke.config`](fluke.config.md): contains classes for configuration;
 - [`fluke.server`](fluke.server.md): contains classes for server-side functionalities;
 - [`fluke.comm`](fluke.comm.md): contains classes for communication;
 - [`fluke.nets`](fluke.nets.md): contains classes for neural networks;
@@ -159,9 +160,7 @@ In this page you can find the list of modules/submodules defined in `fluke` with
 
 .. autosummary::
    :nosignatures:
-   
-   OptimizerConfigurator
-   Configuration
+
    ClientObserver
    ServerObserver
