+++ docs/fluke.config.md
@@ -1,0 +1,64 @@
+(fluke.config)=
+
+# ``fluke.config``
+
+```{eval-rst}
+
+.. automodule:: fluke.config
+   :no-members:
+
+```
+
+
+```{eval-rst}
+
+.. autosummary:: 
+   :nosignatures:
+
+   Configuration
+   ConfigurationError
+   OptimizerConfigurator
+
+```
+
+<h3>
+
+{bdg-primary}`class` ``fluke.config.Configuration``
+
+</h3>
+
+```{eval-rst}
+
+.. autoclass:: fluke.config.Configuration
+   :members:
+
+```
+
+<h3>
+
+{bdg-primary}`class` ``fluke.config.ConfigurationError``
+
+</h3>
+
+```{eval-rst}
+
+.. autoclass:: fluke.config.ConfigurationError
+   :members:
+
+```
+
+
+<h3>
+
+{bdg-primary}`class` ``fluke.config.OptimizerConfigurator``
+
+</h3>
+
+```{eval-rst}
+
+.. autoclass:: fluke.config.OptimizerConfigurator
+   :members: __call__
+
+```
+
+<h3>
