+++ fluke/algorithms/fedala.py
@@ -8,7 +8,6 @@ References:
 """
 import sys
 from copy import deepcopy
-from typing import Any
 
 import numpy as np
 import torch
@@ -18,9 +17,9 @@ sys.path.append(".")
 sys.path.append("..")
 
 from ..client import Client  # NOQA
+from ..config import OptimizerConfigurator  # NOQA
 from ..data import FastDataLoader  # NOQA
 from ..nets import EncoderHeadNet  # NOQA
-from ..utils import OptimizerConfigurator  # NOQA
 from ..utils.model import safe_load_state_dict  # NOQA
 from . import CentralizedFL  # NOQA
