+++ fluke/algorithms/fedrod.py
@@ -7,7 +7,7 @@ References:
 """
 import sys
 from copy import deepcopy
-from typing import Any, Literal
+from typing import Literal
 
 import numpy as np
 import torch
@@ -19,10 +19,11 @@ sys.path.append(".")
 sys.path.append("..")
 
 from ..client import Client  # NOQA
+from ..config import OptimizerConfigurator  # NOQA
 from ..data import FastDataLoader  # NOQA
 from ..evaluation import Evaluator  # NOQA
 from ..nets import EncoderHeadNet  # NOQA
-from ..utils import OptimizerConfigurator, clear_cuda_cache  # NOQA
+from ..utils import clear_cuda_cache  # NOQA
 from ..utils.model import ModOpt  # NOQA
 from . import CentralizedFL  # NOQA
