+++ tests/test_server.py
@@ -13,10 +13,11 @@ from fluke import DDict  # NOQA
 from fluke import FlukeENV  # NOQA
 from fluke.client import Client  # NOQA
 from fluke.comm import Channel  # NOQA
+from fluke.config import OptimizerConfigurator  # NOQA
 from fluke.data import FastDataLoader  # NOQA
 from fluke.evaluation import ClassificationEval  # NOQA
 from fluke.server import Server  # NOQA
-from fluke.utils import OptimizerConfigurator, ServerObserver  # NOQA
+from fluke.utils import ServerObserver  # NOQA
 
 
 def test_server():
